Fix JSON test failure
➜  go-dnsimple git:(register) ✗ go test ./dnsimple
body: float64
want: int
--- FAIL: TestRegistrarService_Register (0.00 seconds)
dnsimple_test.go:58: Request parameters =
map[domain:map[registrant_id:21 name:example.com]], want
map[domain:map[name:example.com registrant_id:21]]
FAIL
FAIL	github.com/weppos/go-dnsimple/dnsimple	0.019s

See https://groups.google.com/forum/#!topic/golang-nuts/TDuGDJAIuVM

+++ dnsimple/registrar_test.go
@@ -77,7 +77,7 @@ func TestRegistrarService_Register(t *testing.T) {
 
 	mux.HandleFunc("/v1/domain_registrations", func(w http.ResponseWriter, r *http.Request) {
 		want := make(map[string]interface{})
-		want["domain"] = map[string]interface{}{"name": "example.com", "registrant_id": 21}
+		want["domain"] = map[string]interface{}{"name": "example.com", "registrant_id": float64(21)}
 
 		testMethod(t, r, "POST")
 		testRequestJSON(t, r, want)
